Create README.md
Analysis of Restaurants DatasetAnalysis of Restaurants Dataset
Jul 2023 - Jul 2023Jul 2023 - Jul 2023

Associated with Cognifyz TechnologiesAssociated with Cognifyz Technologies
The objective of this project is to analyze a restaurants dataset and derive meaningful insights from it. The dataset provides information about various restaurants, including details about cuisines served and their respective locations. The key insights we aim to uncover include identifying the top three most common cuisines, calculating the percentage of restaurants serving each of these cuisines, and determining the city with the highest number of restaurants in the dataset.

To achieve our goals, we followed a systematic approach. We first loaded and preprocessed the dataset to ensure data quality and consistency. Then, we performed exploratory data analysis to gain initial insights into the dataset's structure and content.

Insights are:-
Determine the top three most common cuisines in the dataset.
Calculate the percentage of restaurants that serve each of the top cuisines.
Identify the city with the highest number of restaurants in the dataset.
Calculate the average rating for restaurants in each city.
Determine the city with the highest average rating.
Create a histogram or bar chart to visualize the distribution of price ranges among the restaurants.
Calculate the percentage of restaurants in each price range category.
Determine the percentage of restaurants that offer online delivery.
Compare the average ratings of restaurants with and without online delivery.
Analyze the distribution of aggregate ratings and determine the most common rating range.
Calculate the average number of votes received by restaurants.
Identify the most common combinations of cuisines in the dataset.
Determine if certain cuisine combinations tend to have higher ratings.
and many more...